move collector / http_context test in to separate file
